requirements.txt
torch>=2.0.0+cu118
numpy>=1.23.0
torchvision>=0.15.1+cu118
pillow>=9.5.0
scikit-learn>=1.3.0
opencv-python>=3.4.15.55
scipy>=1.10.1
pandas>=2.0.0
scikit-image>=0.21.0
tqdm>=4.65.0